Added upsample option and some debugging plots
Added an option to upsample edge coordinates.  This was part of some experimentation
to see if I could increase the visible additive-width overlap of edges that overlapped
along portions between the control points, in an attempt to eliminate the bubbles-on-a-
string look that occassionally ocurrs when two lines overlap.  Didn't solve the problem
robustly (sometimes worked, sometimes made it worse), but I'll leave the option in
there for possible use down the road.

I also added a few plots that only show up if a debugging flag is turned on
(it's hard-coded false right now).